회비관리 수정

+++ src/main/java/com/baemin/repositories/InterviewRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.baemin.domain.entity.Interview;
 
-public interface InterviewRepository extends JpaRepository<Interview, Long> {
+public interface InterviewRepository extends JpaRepository<Interview, Integer> {
 
 }
